# MS-930 Implement stateful face matcher that reduces amount of allocations