--- a/face/infra/bio-sdk-resolver/src/main/java/com/simprints/face/infra/biosdkresolver/RocV1BioSdk.kt
+++ b/face/infra/bio-sdk-resolver/src/main/java/com/simprints/face/infra/biosdkresolver/RocV1BioSdk.kt
@@ -1,6 +1,9 @@
 package com.simprints.face.infra.biosdkresolver
 
+import com.simprints.face.infra.basebiosdk.matching.FaceMatcher
+import com.simprints.face.infra.basebiosdk.matching.FaceSample
 import com.simprints.face.infra.rocv1.detection.RocV1Detector
+import com.simprints.face.infra.rocv1.detection.RocV1Detector.Companion.RANK_ONE_TEMPLATE_FORMAT_1_23
 import com.simprints.face.infra.rocv1.initialization.RocV1Initializer
 import com.simprints.face.infra.rocv1.matching.RocV1Matcher
 import javax.inject.Inject
@@ -10,7 +13,10 @@ import javax.inject.Singleton
 class RocV1BioSdk @Inject constructor(
     override val initializer: RocV1Initializer,
     override val detector: RocV1Detector,
-    override val matcher: RocV1Matcher,
 ) : FaceBioSDK {
     override val version: String = "1.23"
+    override val templateFormat: String = RANK_ONE_TEMPLATE_FORMAT_1_23
+    override val matcherName: String = "RANK_ONE"
+
+    override fun createMatcher(probeSamples: List<FaceSample>): FaceMatcher = RocV1Matcher(probeSamples)
 }