--- a/face/infra/bio-sdk-resolver/src/main/java/com/simprints/face/infra/biosdkresolver/RocV3BioSdk.kt
+++ b/face/infra/bio-sdk-resolver/src/main/java/com/simprints/face/infra/biosdkresolver/RocV3BioSdk.kt
@@ -1,6 +1,9 @@
 package com.simprints.face.infra.biosdkresolver
 
+import com.simprints.face.infra.basebiosdk.matching.FaceMatcher
+import com.simprints.face.infra.basebiosdk.matching.FaceSample
 import com.simprints.face.infra.rocv3.detection.RocV3Detector
+import com.simprints.face.infra.rocv3.detection.RocV3Detector.Companion.RANK_ONE_TEMPLATE_FORMAT_3_1
 import com.simprints.face.infra.rocv3.initialization.RocV3Initializer
 import com.simprints.face.infra.rocv3.matching.RocV3Matcher
 import javax.inject.Inject
@@ -10,7 +13,10 @@ import javax.inject.Singleton
 class RocV3BioSdk @Inject constructor(
     override val initializer: RocV3Initializer,
     override val detector: RocV3Detector,
-    override val matcher: RocV3Matcher,
 ) : FaceBioSDK {
     override val version: String = "3.1"
+    override val templateFormat: String = RANK_ONE_TEMPLATE_FORMAT_3_1
+    override val matcherName: String = "RANK_ONE"
+
+    override fun createMatcher(probeSamples: List<FaceSample>): FaceMatcher = RocV3Matcher(probeSamples)
 }