--- a/face/infra/bio-sdk-resolver/src/test/java/com/simprints/face/infra/biosdkresolver/RocV1BioSdkTest.kt
+++ b/face/infra/bio-sdk-resolver/src/test/java/com/simprints/face/infra/biosdkresolver/RocV1BioSdkTest.kt
@@ -1,0 +1,22 @@
+package com.simprints.face.infra.biosdkresolver
+
+
+import com.google.common.truth.Truth.*
+import io.mockk.*
+import org.junit.Test
+
+class RocV1BioSdkTest {
+
+
+    private lateinit var rocV1BioSdk: RocV1BioSdk
+
+    @Test
+    fun createMatcher() {
+        rocV1BioSdk = RocV1BioSdk(mockk(), mockk())
+
+        val matcher = rocV1BioSdk.createMatcher(emptyList())
+
+        assertThat(matcher).isNotNull()
+    }
+
+}