--- a/face/infra/bio-sdk-resolver/src/test/java/com/simprints/face/infra/biosdkresolver/RocV3BioSdkTest.kt
+++ b/face/infra/bio-sdk-resolver/src/test/java/com/simprints/face/infra/biosdkresolver/RocV3BioSdkTest.kt
@@ -1,0 +1,21 @@
+package com.simprints.face.infra.biosdkresolver
+
+
+import com.google.common.truth.Truth.*
+import io.mockk.*
+import org.junit.Test
+
+class RocV3BioSdkTest {
+
+    private lateinit var rocV3BioSdk: RocV3BioSdk
+
+    @Test
+    fun createMatcher() {
+        rocV3BioSdk = RocV3BioSdk(mockk(), mockk())
+
+        val matcher = rocV3BioSdk.createMatcher(emptyList())
+
+        assertThat(matcher).isNotNull()
+    }
+
+}