--- a/face/infra/roc-v1/src/main/java/com/simprints/face/infra/rocv1/RocV1WrapperModule.kt
+++ b/face/infra/roc-v1/src/main/java/com/simprints/face/infra/rocv1/RocV1WrapperModule.kt
@@ -5,7 +5,6 @@ import com.simprints.face.infra.basebiosdk.initialization.FaceBioSdkInitializer
 import com.simprints.face.infra.basebiosdk.matching.FaceMatcher
 import com.simprints.face.infra.rocv1.detection.RocV1Detector
 import com.simprints.face.infra.rocv1.initialization.RocV1Initializer
-import com.simprints.face.infra.rocv1.matching.RocV1Matcher
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -19,7 +18,4 @@ abstract class RocV1WrapperModule {
 
     @Binds
     abstract fun provideFaceDetector(impl: RocV1Detector): FaceDetector
-
-    @Binds
-    abstract fun provideFaceMatcher(impl: RocV1Matcher): FaceMatcher
 }