--- a/face/infra/roc-v1/src/test/java/com/simprints/infra/rocwrapper/matching/RocV1MatcherTest.kt
+++ b/face/infra/roc-v1/src/test/java/com/simprints/infra/rocwrapper/matching/RocV1MatcherTest.kt
@@ -1,6 +1,6 @@
 package com.simprints.infra.rocwrapper.matching
 
-import com.google.common.truth.Truth
+import com.google.common.truth.*
 import com.simprints.face.infra.rocv1.matching.RocV1Matcher
 import org.junit.Test
 
@@ -8,7 +8,6 @@ class RocV1MatcherTest {
     // Dummy test to generate jacoco reports.
     @Test
     fun getMatcherName() {
-        RocV1Matcher().matcherName
-        Truth.assertThat(RocV1Matcher().matcherName).isEqualTo("RANK_ONE")
+        Truth.assertThat(RocV1Matcher(emptyList())).isNotNull()
     }
 }