--- a/face/infra/roc-v3/src/test/java/com/simprints/infra/rocwrapper/matching/RocV3MatcherTest.kt
+++ b/face/infra/roc-v3/src/test/java/com/simprints/infra/rocwrapper/matching/RocV3MatcherTest.kt
@@ -4,10 +4,10 @@ import com.google.common.truth.Truth
 import com.simprints.face.infra.rocv3.matching.RocV3Matcher
 import org.junit.Test
 
+// Dummy test to generate jacoco reports.
 class RocV3MatcherTest {
     @Test
     fun getMatcherName() {
-        RocV3Matcher().matcherName
-        Truth.assertThat(RocV3Matcher().matcherName).isEqualTo("RANK_ONE")
+        Truth.assertThat(RocV3Matcher(emptyList())).isNotNull()
     }
 }