[JIRA ticket](https://simprints.atlassian.net/browse/MS-930)
Will be released in: **2025.2.0**

This is an experimental implementation of the suggestion in the JIRA ticket, it could be polished to be production-ready if required.

### Notable changes

* Changing the face matches to pre-allocate native sample template memory and reuse it for each candidate. In theory, it should reduce memory and GC pressure.
* Still using a separate allocation per coroutine to avoid cross-coroutine memory sharing and potential concurrency issues. There is still a significant reduction in allocation calls.

Surprisingly the change is way less noticeable in my very basic benchmarking (running identification on 50k candidates using ROCv3 on a Lenovo Tab 8): 

| Before change  | After change |
|--------|--------|
| 114780ms | 92866ms  |
| 113032ms | 110177ms |
| 107323ms | 110511ms |
| 110483ms | 95675ms  |
| 111383ms | 104624ms |
| 113065ms | 109125ms |
| 112612ms | 110679ms |
| Avg: **111811** | Avg: **104808**


As part of this task, I also investigated the feasibility of the same refactoring for fingerprints and decided against it for the following reasons:

* SimAfis matching supports 1:N on the library level and takes the candidate list as a parameter, therefore it is highly likely that the same allocation optimisation is happening under the hood (or at least it should be done on the library level).
* NEC wraps templates with a custom data class and there is no way to control allocations. Therefore even if we cache the template list, it would not prevent unnecessary allocations.
* Fingerprint BioSDK setup is significantly more complex and would require a lot of changes for no/negligible benefit.

### Additional work checklist

* [x] Effect on other features and security has been considered
* [x] Design document marked as "In development" (if applicable)
* [x] External (Gitbook) and internal (Confluence) Documentation is up to date (or ticket created)
* [x] Test cases in Testiny are up to date (or ticket created)
* [x] Other teams notified about the changes (if applicable)